+++ lsd-slam-pangolin/Tracking/Relocalizer.cpp
@@ -76,14 +76,9 @@ void Relocalizer::updateCurrentFrame(std::shared_ptr<Frame> currentFrame)
 	if(hasResult) return;
 
 	this->CurrentRelocFrame = currentFrame;
-//	int doneLast = KFForReloc.size() - (maxRelocIDX-nextRelocIDX);
-	maxRelocIDX = nextRelocIDX + KFForReloc.size();
+	maxRelocIDX = (int) (nextRelocIDX + KFForReloc.size());
 	newCurrentFrameSignal.notify_all();
 	lock.unlock();
-
-//	printf("tried last on %d. set new current frame %d. trying %d to %d!\n",
-//			doneLast,
-//			currentFrame->id(), nextRelocIDX, maxRelocIDX);
 
 	if (displayDepthMap)
 		Util::displayImage( "DebugWindow DEPTH", cv::Mat(currentFrame->height(), currentFrame->width(), CV_32F, currentFrame->image())*(1/255.0f), false );
@@ -91,6 +86,7 @@ void Relocalizer::updateCurrentFrame(std::shared_ptr<Frame> currentFrame)
 	int pressedKey = Util::waitKey(1);
 	handleKey(pressedKey);
 }
+
 void Relocalizer::start(std::vector<Frame*, Eigen::aligned_allocator<lsd_slam::Frame*> > &allKeyframesList)
 {
 	// make KFForReloc List
